FEAT: Update atomic orderings for job latch to use Release..Acquire
The usage here is a typical Release..Acquire scenario; data is written
on the job executing thread, then we store to the atomic with Release.

When loading with Acquire in the waiting thread, all the writes in the
executing thread before the store are visible as *side effects* on the
waiting thread.